Add showcase and gen9 badges to player table
Gen9 is not enabled yet, but will probably be in september.
Showcase badge seems to only be for max size showcases. They might add more badges later.
Route badges that were unused are still unused now. So I removed the comments.
Route stats are in another proto which isn't available in the public profile.