Support PogoDroid lat/lng/quests_held
PogoDroid sends "quests_held" as a list with quest_ids, instead of
"have_ar". This list can be scanned for an AR task to come up with
"have_ar".

Additionally, the parsing of the lat/lng from PogoDroid was missing.
This adds the lat/lng parsing, which fixes the device location tracking
such that PD devices will now show up in ReactMap.