Do not copy logs and cache into docker image
The cache/logs shouldn't be copied back into new docker image builds,
which is what happens when you attach these as volumes, run Golbat
for a while, and then re-build the image. This was shoving ~175M
of my logs to the docker daemon during build and copying into the
image!

I think it only makes sense for people to attach these as volumes
at runtime, and doing anything with them at image build time could
just end up being an expensive build time no-op.

But even if not, if people do not attach these as volumes, they are
always going to be empty on the host side such that there's nothing
to copy into the image. But we still need to create the directories.

If you've compiled outside of docker, this also stops the sending of
the 130M golbat binary through to the docker daemon when you next
build, also.