Handle larger raw bodies
PD is sending the game master proto and the size exceeds the 1M
limit. I was seeing close to 2.5M for it. Since this is only to guard
against something having gone haywire or a DoS if you have this wide
open, it is rather safe to bump this to 5M.