Allow docker build to make use of vendor
Having to download dependencies every time during a docker build
after a code change makes the build take longer than if
dependencies were simply vendored.

This just adds a check to the Dockerfile to see if vendor/ is
being used, and if so, skips downloading.

After this lands, one can manage keeping vendor/ up to date in
their workspace via 'go mod vendor' and a docker build will not
need to download anything. This reduces the docker build time by
~25% for me, but YMMV.