Raise max pokemon id
As it is, it looks like we would crash if we tried to count
stats for Goldhengo (dex #1000) or anything later. Paldea
currently takes us up to 1017, but there are at least 4 more
that will be in the 2nd part of the Scarlet/Violet DLC.

So, this PR raises the max pokedex number to 1050 to be safe.