Accurate shiny stats tracking
This PR computes shiny stats, but does so by the combination
of 'encounterId,accountUsername'. We need to look at every
unique encounterId/accountUsername pair, so this is hooked
in before we check if an encounter is a duplicate, and only
in the encounter paths.

This converts the existing 'pokemonTimings' cache which is
keyed by encounterId into a more general encounter cache,
adding the map of accounts that have seen the encounter to it.

Shiny stats are not computed per area. After discussion, we
determined that area is not very useful and it's too expensive
to compute just for these stats. There can be a lot of duplicate
encounters, and this avoids the work to compute the areas on
duplicates.

For the DB/in-memory stats, they will be updated with area
'world'.

For prometheus, this sends shiny,shundo,snundo stats. For
'any shiny', I'm sending metrics by dex+form Id. And non-shiny
counts are submitted as well by dex+form Id. The purpose for this
is so that we can compute accurate shiny odds per dex/form.

A metric is also added to count the number of times the same
encounterId is seen for both same and different accounts.