Add DB call count stats
I was trying to track down some IO spikes to my DB filesystem. Strangely
nothing really turned up in slow query log. So I thought maybe there was
a spam of requests happening for some strange reason. So, to check or
eliminate golbat, I added these stats. (It turns out this eliminated
Golbat and made me look elsewhere for the stupid thing I was doing.)

This might be interesting for others, also, so here it is. And timings
might also be useful to add at some point to catch degradation.