Bump go requiremment to 1.22
This bumps golang requirement to 1.22.

1.22.1 is out now and while it doesn't appear that
we have any workaround for 'for loop variable scope' in Golbat
(unless my search failed me)...it would be nice to not have to
worry about it going forward.

There's also a new sql.Null generic. And also 'for i := range <integer>'
to range from 0 to <integer>-1.

Supposedly memory performance is improved, also.