experimental: in-memory forts
adds full support for keeping a rtree of forts in-memory and querying from it on an endpoint, similar to how pokemon work.

* enable `fort_in_memory` in config
* initial filling up the rtree takes a while
* fetching forts that are not cached will query them one-by-one, leading to long response times (even after filtering was near-instant)

ideally, there should be a fallback to query all pokestops for when this feature is turned off or while it's still loading up the rtree.

also, while fetching it should query all missing pokestops at once, not one-by-one.